% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRCovFiltration.R
\name{GRCovFiltration}
\alias{GRCovFiltration}
\title{Exclusion of restriction sites with insufficient coverage}
\usage{
GRCovFiltration(
  bamFiles,
  GR,
  tread = 1,
  mincov = 30,
  minSample = 1,
  duplicate = T
)
}
\arguments{
\item{bamFiles}{Vector with bam file names}

\item{GR}{GRanges object with coordinates of restriction sites}

\item{tread}{The number of running threads at the same time, doesn't work in windows.}

\item{mincov}{Minimum coverage.}

\item{minSample}{Minimum number of samples for which the restriction site has the minimum amount of coverage.}

\item{duplicate}{Remove duplicates}
}
\value{
Filtered Granges object
}
\description{
Exclusion of restriction sites with insufficient coverage
}
