% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metKitExport.R
\name{metKitExport}
\alias{metKitExport}
\title{Export results in a format compatible with methylKit}
\usage{
metKitExport(bamFiles, GR, mincov = 30, save_dir = "./", duplicate = T)
}
\arguments{
\item{bamFiles}{Vector with bam file names}

\item{GR}{GRanges object with coordinates of restriction sites}

\item{mincov}{Minimum coverage.}

\item{save_dir}{Directory for saving results.}

\item{duplicate}{Remove duplicates}
}
\description{
Export results in a format compatible with methylKit
}
