% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMetSeq.R
\name{RMetSeq}
\alias{RMetSeq}
\title{Estimation of the level of methylation}
\usage{
RMetSeq(bamFiles, GR, tread = 1, duplicate = T)
}
\arguments{
\item{bamFiles}{Vector with bam file names}

\item{GR}{GRanges object with coordinates of restriction sites}

\item{tread}{The number of running threads at the same time, doesn't work in windows.}

\item{duplicate}{Remove duplicates}
}
\value{
Data frame with the level of methylation and coverage by restriction sites
}
\description{
Estimation of the level of methylation
}
